LayoutTests/js/arrowfunction-asparamter-1-expected.txt

Tests for ES6 arrow function, passing arrow function as the paramter

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" + [1, 2, 3, 4].map(x => x, 32) is '1,2,3,4'
PASS successfullyParsed is true

TEST COMPLETE
